feat: Enable module support for Babel
This change adds the `data-type="module"` attribute to the Babel script tag in `index.html`. This allows the application to leverage ES module syntax in the `index.tsx` file, enabling better code organization and reusability.

The removal of the empty line at the end of `GeneratedScheduleView.tsx` is a minor cleanup.